api/requirements.txt
Cython==0.25.2
Flask==0.12
Flask-Assets==0.12
Flask-Cache==0.13.1
Flask-Cors==3.0.2
Flask-ErrorMail==0.2.2
Flask-HTTPAuth==3.2.2
Flask-KVSession==0.6.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.2
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-WTF==0.14.2
Flask-Webpack==0.1.0
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==0.23
PyJWT==1.5.3
PyMySQL==0.7.10
PyYAML==3.12
Pygments==2.2.0
SQLAlchemy==1.1.5
WTForms==2.1
Werkzeug==0.11.15
alembic==0.8.10
amqp==2.1.4
aniso8601==1.2.0
anyjson==0.3.3
appdirs==1.4.0
arrow==0.10.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.3
billiard==3.5.0.2
blinker==1.4
boto3==1.4.4
botocore==1.5.14
bs4==0.0.1
celery==4.0.2
certifi==2017.1.23
cffi==1.9.1
click==6.7
cryptography==1.7.2
decorator==4.0.11
docutils==0.13.1
enum34==1.1.6
et-xmlfile==1.0.1
futures==3.0.5
idna==2.2
ipaddress==1.0.18
ipython==5.3.0
ipython-genutils==0.1.0
itsdangerous==0.24
jmespath==0.9.1
jwt==0.5.2
kombu==4.0.2
lockfile==0.12.2
marshmallow==2.13.0
packaging==16.8
passlib==1.6.2
pathlib2==2.2.1
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.2
pycparser==2.17
pyparsing==2.2.0
python-daemon==2.1.2
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
redis==2.10.5
requests==2.13.0
rsa==3.4.2
s3transfer==0.1.10
scandir==1.5
simplegeneric==0.8.1
simplejson==3.10.0
simplekv==0.10.0
singledispatch==3.4.0.3
six==1.10.0
suds==0.4
suds-jurko==0.6
traitlets==4.3.1
typing==3.5.3.0
vine==1.1.3
wcwidth==0.1.7
webassets==0.12.1
xlrd==1.0.0
xmltodict==0.10.2
